% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{btwExtra_default_tools}
\alias{btwExtra_default_tools}
\alias{btwExtra_mcp_server}
\title{Combined tools for MCP with btw and btwExtra}
\usage{
btwExtra_default_tools()

btwExtra_mcp_server(tools = btwExtra_default_tools())
}
\arguments{
\item{tools}{List of \code{ellmer::tool()} or a path to an R script compatible
with \code{mcptools::mcp_server()}. By default it uses \code{btwExtra_default_tools()}.}
}
\value{
List of tools (\code{btwExtra_default_tools()}) or the result of
\code{btw::btw_mcp_server()} (\code{btwExtra_mcp_server()}).
}
\description{
\code{btwExtra_default_tools()} returns the MCP-friendly combination: all
\code{btw::btw_tools()} plus the additions from \code{btwExtra_tools()}.

\code{btwExtra_mcp_server()} runs \code{btw::btw_mcp_server()} using that default set
(or a custom set if provided).
}
