% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-run-r-code.R
\name{btwExtra_tool_env_run_r_code}
\alias{btwExtra_tool_env_run_r_code}
\title{Tool: Run arbitrary R code}
\usage{
btwExtra_tool_env_run_r_code(code, max_output_lines = 20L, `_intent` = "")
}
\arguments{
\item{code}{String with the R code to execute.}

\item{max_output_lines}{Maximum number of output lines to display; use \code{-1}
to disable truncation. Defaults to 20.}
}
\description{
Escape hatch to evaluate R code in the connected session, emulating the R
console: each expression is evaluated in order, assignments are not printed,
and any visible value is shown. Console output is captured and can be
truncated to save context.
}
